pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-trading-system"
version = "0.1.0"
description = "An autonomous, modular trading system using Reinforcement Learning."
requires-python = ">=3.9"
dependencies = [
    "python-dotenv==1.0.1",
    "psycopg-pool==3.2.0",
    "requests",
    "backoff",
    "apscheduler",
    "python-telegram-bot",
    "flask",
    "healthcheck",
    "streamlit",
    "pandas==2.2.2",
    "numpy>=1.26.0,<2.0",
    "tensortrade",
    "stable-baselines3==2.7.0",
    "yfinance==0.2.37",
    "torch",
    "deepxde",
    "gymnasium==0.29.1",
    "shimmy==0.2.1"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "black",
    "isort",
]

[tool.setuptools.packages.find]
where = ["src"]
